# docs/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customer/gifts/{id}": {
      "get": {
        "description": "REGISTERED_USER only. Hidden/admin-only gift records are never returned.",
        "operationId": "CustomerMarketplaceController_giftDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gift details fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch customer-safe gift details",
        "tags": [
          "Customer Marketplace"
        ]
      }
    },
...
    "/api/v1/gifts": {
      "post": {
        "operationId": "GiftsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGiftDto"
              },
              "examples": {
                "withVariants": {
                  "value": {
                    "name": "Luxury Perfume",
                    "description": "Long-lasting premium fragrance.",
                    "shortDescription": "Premium fragrance gift.",
                    "categoryId": "gift_category_id",
                    "providerId": "provider_id",
                    "price": 99.99,
                    "currency": "PKR",
                    "stockQuantity": 50,
                    "sku": "PERFUME-001",
                    "imageUrls": [
                      "https://cdn.yourdomain.com/gift-images/perfume.png"
                    ],
                    "isPublished": true,
                    "isFeatured": false,
                    "tags": [
                      "perfume",
                      "luxury"
                    ],
                    "moderationStatus": "APPROVED",
                    "variants": [
                      {
                        "name": "30ml",
                        "price": 89.99,
                        "originalPrice": 119.99,
                        "stockQuantity": 10,
                        "sku": "PERFUME-30ML",
                        "isPopular": false,
                        "isDefault": false,
                        "sortOrder": 1,
                        "isActive": true
                      },
                      {
                        "name": "50ml",
                        "price": 129.99,
                        "originalPrice": 159.99,
                        "stockQuantity": 20,
                        "sku": "PERFUME-50ML",
                        "isPopular": true,
                        "isDefault": true,
                        "sortOrder": 2,
                        "isActive": true
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      },
      "get": {
        "operationId": "GiftsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "ACTIVE",
                "INACTIVE",
                "PENDING",
                "REJECTED",
                "FLAGGED",
                "OUT_OF_STOCK"
              ]
            }
          },
          {
            "name": "moderationStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "PENDING",
                "APPROVED",
                "REJECTED",
                "FLAGGED"
              ]
            }
          },
          {
            "name": "isPublished",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "price",
                "rating",
                "stockQuantity"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      }
    },
    "/api/v1/gifts/stats": {
      "get": {
        "operationId": "GiftsController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      }
    },
    "/api/v1/gifts/export": {
      "get": {
        "operationId": "GiftsController_export",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "ACTIVE",
                "INACTIVE",
                "PENDING",
                "REJECTED",
                "FLAGGED",
                "OUT_OF_STOCK"
              ]
            }
          },
          {
            "name": "moderationStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "PENDING",
                "APPROVED",
                "REJECTED",
                "FLAGGED"
              ]
            }
          },
          {
            "name": "isPublished",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "price",
                "rating",
                "stockQuantity"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CSV",
                "XLSX"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      }
    },
    "/api/v1/gifts/{id}": {
      "get": {
        "operationId": "GiftsController_details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      },
      "patch": {
        "operationId": "GiftsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGiftDto"
              },
              "examples": {
                "upsertVariants": {
                  "value": {
                    "name": "Luxury Perfume Updated",
                    "replaceVariants": false,
                    "variants": [
                      {
                        "id": "variant_id",
                        "name": "50ml",
                        "price": 129.99,
                        "originalPrice": 159.99,
                        "stockQuantity": 20,
                        "sku": "PERFUME-50ML",
                        "isPopular": true,
                        "isDefault": true,
                        "sortOrder": 2,
                        "isActive": true
                      },
                      {
                        "name": "150ml",
                        "price": 249.99,
                        "originalPrice": 299.99,
                        "stockQuantity": 5,
                        "sku": "PERFUME-150ML",
                        "isPopular": false,
                        "isDefault": false,
                        "sortOrder": 4,
                        "isActive": true
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      },
      "delete": {
        "operationId": "GiftsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      }
    },
    "/api/v1/gifts/{id}/status": {
      "patch": {
        "operationId": "GiftsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGiftStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gift Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gift App Backend API",
    "description": "Gift App authentication and user account APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GiftVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "variant_id"
          },
          "name": {
            "type": "string",
            "example": "50ml"
          },
          "price": {
            "type": "number",
            "example": 129.99
          },
          "originalPrice": {
            "type": "number",
            "example": 159.99
          },
          "stockQuantity": {
            "type": "number",
            "example": 20
          },
          "sku": {
            "type": "string",
            "example": "PERFUME-50ML"
          },
          "isPopular": {
            "type": "boolean",
            "example": true
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 2
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "price",
          "stockQuantity"
        ]
      },
      "CreateGiftDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "stockQuantity": {
            "type": "number"
          },
          "sku": {
            "type": "string"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "moderationStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "FLAGGED"
            ]
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GiftVariantDto"
            }
          }
        },
        "required": [
          "name",
          "categoryId",
          "providerId",
          "price"
        ]
      },
      "UpdateGiftDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "stockQuantity": {
            "type": "number"
          },
          "sku": {
            "type": "string"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "replaceVariants": {
            "type": "boolean",
            "example": false
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GiftVariantDto"
            }
          }
        }
      },
      "UpdateGiftStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "OUT_OF_STOCK"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}